requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==1.45.0
lingua-language-detector==2.0.2
pydantic==2.9.0
python-dotenv==1.0.1
